refactor: handler subpackage
