+++ api/dto/arrayresponse.dto.go
@@ -1,5 +1,1 @@
 package dto
-
-type ArrayResponse[T any] struct {
-	Data []*T `json:"data"`
-}
