+++ api/infrastructure/handler/auth/login.go
@@ -1,0 +1,49 @@
+package auth
+
+import (
+	"context"
+	"net/http"
+
+	"github.com/danielgtaylor/huma/v2"
+	"shanepee.com/api/domain"
+	"shanepee.com/api/infrastructure/handler"
+)
+
+type LoginBody struct {
+	Email    string `json:"email" example:"johndoe@example.com"`
+	Password string `json:"password" example:"VerySecurePassword"`
+}
+
+type LoginInput struct {
+	Body LoginBody
+}
+
+type LoginOutput struct {
+	Body *domain.User
+}
+
+func (h *AuthHandler) RegisterLogin(api huma.API) {
+	huma.Register(api, huma.Operation{
+		OperationID: "login",
+		Method:      http.MethodPost,
+		Path:        "/v1/auth/login",
+		Tags:        []string{"Authentication"},
+		Summary:     "Login User",
+		Description: "Login",
+	}, func(ctx context.Context, i *LoginInput) (*LoginOutput, error) {
+		data, err := h.authSvc.Login(ctx, i.Body.Email, i.Body.Password)
+		if err != nil {
+			return nil, handler.ErrForbidden
+		}
+
+		session := handler.GetSession(ctx)
+		session.Set(handler.UserIdSessionKey, data.ID)
+		if err := session.Save(); err != nil {
+			return nil, handler.ErrIntervalServerError
+		}
+
+		return &LoginOutput{
+			Body: data,
+		}, nil
+	})
+}
